password reset
Implemented password reset functionality via sending a recovery code
