+++ helpers/generateRandomCode.js
@@ -1,0 +1,12 @@
+function generateRandomCode(length) {
+    const characters =
+      "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz";
+    let code = "";
+    for (let i = 0; i < length; i++) {
+      const randomIndex = Math.floor(Math.random() * characters.length);
+      code += characters.charAt(randomIndex);
+    }
+    return code;
+}
+
+export default generateRandomCode
