+++ helpers/index.js
@@ -1,3 +1,4 @@
 export {default as HttpError} from './HttpError.js'
 export {default as sendMail} from './sendMail.js'
 export {default as cloudinary} from './cloudinary.js';
+export {default as generateRandomCode} from './generateRandomCode.js';
