+++ routes/auth.js
@@ -13,7 +13,8 @@ authRouter.post("/signout", authenticate, authController.signout);
 authRouter.get("/current", authenticate, authController.getCurrent);
 authRouter.patch("/edit", authenticate, upload.single('avatar'), resizeAvatar, isEmptyBody, validateBody(userEditScheme), authController.editProfile);
 authRouter.post("/needHelp", authenticate, isEmptyBody, validateBody(userHelpMailScheme), authController.sendNeedHelp);
-authRouter.post("/recovery", authenticate, isEmptyBody, authController.forgotPassword)
+authRouter.post("/recovery-mail", authenticate, isEmptyBody, authController.forgotPassword)
+authRouter.patch("/reset-password", isEmptyBody, authController.resetPassword)
 
 // authRouter.get("/verify/:verificationToken")
 // authRouter.post("/verify", isEmptyBody, validateBody(userEmailScheme))
